+++ src/jpegls.cpp
@@ -54,33 +54,33 @@ const std::error_category& CharLSCategoryInstance()
 }
 
 
-signed char QuantizeGratientOrg(const JlsCustomParameters& preset, LONG NEAR, LONG Di)
+signed char QuantizeGratientOrg(const JlsCustomParameters& preset, int32_t NEAR, int32_t Di)
 {
-	if (Di <= -preset.T3) return  -4;
-	if (Di <= -preset.T2) return  -3;
-	if (Di <= -preset.T1) return  -2;
-	if (Di < -NEAR)  return  -1;
-	if (Di <=  NEAR) return   0;
-	if (Di < preset.T1)   return   1;
-	if (Di < preset.T2)   return   2;
-	if (Di < preset.T3)   return   3;
+    if (Di <= -preset.T3) return  -4;
+    if (Di <= -preset.T2) return  -3;
+    if (Di <= -preset.T1) return  -2;
+    if (Di < -NEAR)  return  -1;
+    if (Di <=  NEAR) return   0;
+    if (Di < preset.T1)   return   1;
+    if (Di < preset.T2)   return   2;
+    if (Di < preset.T3)   return   3;
 
-	return  4;
+    return  4;
 }
 
 
-std::vector<signed char> CreateQLutLossless(LONG cbit)
+std::vector<signed char> CreateQLutLossless(int32_t cbit)
 {
-	JlsCustomParameters preset = ComputeDefault((1 << cbit) - 1, 0);
-	LONG range = preset.MAXVAL + 1;
+    JlsCustomParameters preset = ComputeDefault((1 << cbit) - 1, 0);
+    int32_t range = preset.MAXVAL + 1;
 
-	std::vector<signed char> lut(range * 2);
-	
-	for (LONG diff = -range; diff < range; diff++)
-	{
-		lut[range + diff] = QuantizeGratientOrg(preset, 0,diff);
-	}
-	return lut;
+    std::vector<signed char> lut(range * 2);
+    
+    for (int32_t diff = -range; diff < range; diff++)
+    {
+        lut[range + diff] = QuantizeGratientOrg(preset, 0,diff);
+    }
+    return lut;
 }
 
 // Lookup tables to replace code with lookup tables.
@@ -89,9 +89,9 @@ std::vector<signed char> CreateQLutLossless(LONG cbit)
 
 // Lookup table: decode symbols that are smaller or equal to 8 bit (16 tables for each value of k)
 CTable decodingTables[16] = { InitTable(0), InitTable(1), InitTable(2), InitTable(3),
-							  InitTable(4), InitTable(5), InitTable(6), InitTable(7),
-							  InitTable(8), InitTable(9), InitTable(10), InitTable(11),
-							  InitTable(12), InitTable(13), InitTable(14),InitTable(15) };
+                              InitTable(4), InitTable(5), InitTable(6), InitTable(7),
+                              InitTable(8), InitTable(9), InitTable(10), InitTable(11),
+                              InitTable(12), InitTable(13), InitTable(14),InitTable(15) };
 
 
 // Lookup tables: sample differences to bin indexes. 
@@ -104,83 +104,83 @@ std::vector<signed char> rgquant16Ll = CreateQLutLossless(16);
 template<class STRATEGY>
 std::unique_ptr<STRATEGY> JlsCodecFactory<STRATEGY>::GetCodec(const JlsParameters& info, const JlsCustomParameters& presets)
 {
-	std::unique_ptr<STRATEGY> strategy;
+    std::unique_ptr<STRATEGY> strategy;
 
-	if (presets.RESET != 0 && presets.RESET != BASIC_RESET)
-	{
+    if (presets.RESET != 0 && presets.RESET != BASIC_RESET)
+    {
         DefaultTraitsT<uint8_t, uint8_t> traits((1 << info.bitspersample) - 1, info.allowedlossyerror);
-		traits.MAXVAL = presets.MAXVAL;
-		traits.RESET = presets.RESET;
+        traits.MAXVAL = presets.MAXVAL;
+        traits.RESET = presets.RESET;
         strategy = std::unique_ptr<STRATEGY>(new JlsCodec<DefaultTraitsT<uint8_t, uint8_t>, STRATEGY>(traits, info));
-	}
-	else
-	{
-		strategy = GetCodecImpl(info);
-	}
+    }
+    else
+    {
+        strategy = GetCodecImpl(info);
+    }
 
-	if (strategy)
-	{
-		strategy->SetPresets(presets);
-	}
-	return strategy;
+    if (strategy)
+    {
+        strategy->SetPresets(presets);
+    }
+    return strategy;
 }
 
 
 template<class TRAITS, class STRATEGY>
 std::unique_ptr<STRATEGY> CreateCodec(const TRAITS& t, const STRATEGY*, const JlsParameters& info)
 {
-	return std::unique_ptr<STRATEGY>(new JlsCodec<TRAITS, STRATEGY>(t, info));
+    return std::unique_ptr<STRATEGY>(new JlsCodec<TRAITS, STRATEGY>(t, info));
 }
 
 
 template<class STRATEGY>
 std::unique_ptr<STRATEGY> JlsCodecFactory<STRATEGY>::GetCodecImpl(const JlsParameters& info)
 {
-	STRATEGY* s = nullptr;
+    STRATEGY* s = nullptr;
 
-	if (info.ilv == ILV_SAMPLE && info.components != 3)
-		return nullptr;
+    if (info.ilv == ILV_SAMPLE && info.components != 3)
+        return nullptr;
 
 #ifndef DISABLE_SPECIALIZATIONS
 
-	// optimized lossless versions common formats
-	if (info.allowedlossyerror == 0)
-	{
-		if (info.ilv == ILV_SAMPLE)
-		{
-			if (info.bitspersample == 8)
+    // optimized lossless versions common formats
+    if (info.allowedlossyerror == 0)
+    {
+        if (info.ilv == ILV_SAMPLE)
+        {
+            if (info.bitspersample == 8)
                 return CreateCodec(LosslessTraitsT<Triplet<uint8_t>, 8>(), s, info);
-		}
-		else
-		{
-			switch (info.bitspersample)
-			{
+        }
+        else
+        {
+            switch (info.bitspersample)
+            {
                 case  8: return CreateCodec(LosslessTraitsT<uint8_t, 8>(), s, info);
-				case 12: return CreateCodec(LosslessTraitsT<uint16_t, 12>(), s, info);
-				case 16: return CreateCodec(LosslessTraitsT<uint16_t, 16>(), s, info);
-			}
-		}
-	}
+                case 12: return CreateCodec(LosslessTraitsT<uint16_t, 12>(), s, info);
+                case 16: return CreateCodec(LosslessTraitsT<uint16_t, 16>(), s, info);
+            }
+        }
+    }
 
 #endif
 
-	int maxval = (1 << info.bitspersample) - 1;
+    int maxval = (1 << info.bitspersample) - 1;
 
-	if (info.bitspersample <= 8)
-	{
-		if (info.ilv == ILV_SAMPLE)
+    if (info.bitspersample <= 8)
+    {
+        if (info.ilv == ILV_SAMPLE)
             return CreateCodec(DefaultTraitsT<uint8_t, Triplet<uint8_t> >(maxval, info.allowedlossyerror), s, info);
 
         return CreateCodec(DefaultTraitsT<uint8_t, uint8_t>((1 << info.bitspersample) - 1, info.allowedlossyerror), s, info);
-	}
-	else if (info.bitspersample <= 16)
-	{
-		if (info.ilv == ILV_SAMPLE)
-			return CreateCodec(DefaultTraitsT<uint16_t,Triplet<uint16_t> >(maxval, info.allowedlossyerror), s, info);
+    }
+    else if (info.bitspersample <= 16)
+    {
+        if (info.ilv == ILV_SAMPLE)
+            return CreateCodec(DefaultTraitsT<uint16_t,Triplet<uint16_t> >(maxval, info.allowedlossyerror), s, info);
 
-		return CreateCodec(DefaultTraitsT<uint16_t, uint16_t>(maxval, info.allowedlossyerror), s, info);
-	}
-	return nullptr;
+        return CreateCodec(DefaultTraitsT<uint16_t, uint16_t>(maxval, info.allowedlossyerror), s, info);
+    }
+    return nullptr;
 }
